Include `time.h` in `hh_shared.c` since we use `time()`
Summary:
This is only a minor issue, but when I compile Flow on Windows I get

```
3.78s _obuild\hh-heap\hh_shared.o                                                                                                                                               [ done ]
-- stderr for .\_obuild\hh-heap\hh_shared.o --
.\hack\heap\hh_shared.c: In function ‘hh_mem’:
.\hack\heap\hh_shared.c:1756:20: warning: implicit declaration of function ‘time’ [-Wimplicit-function-declaration]
       time_t now = time(0);
```

It's just a warning and everything seems to work fine. But we probably
should include `time.h` if we use `time`. I don't see this warning on
OSX or Linux, since I think `sys/time.h` includes `time.h`.

Reviewed By: samwgoldman

Differential Revision: D7370245

fbshipit-source-id: e59723221584156d0db984e9d3450f76f3fa10c2