Merge low arena and low 1g arena
Summary:
Now that we create the managed arena for low memory even when 1G huge pages are not used, we can retire the old low arena using dss now.
The dss arena is only used when compiling against older versions of jemalloc, which should be deprecated soon.

On builds not using jemalloc, there will be no low arena anyway.

Reviewed By: interwq

Differential Revision: D7161403

fbshipit-source-id: 2a160c81958b0ffc684e9c309cabab15b47a9419