Do not segfault on non-Awaitables in genva()
Summary:
Add input type verifications to the AwaitAll opcode used by the genva()
optimization and move the responsibility for checking nulls to the AwaitAll
opcode.

Without this check, it is possible to segfault HHVM by passing non-Awaitables
to genva(). This did not previously cause issues, as we used to call the
getWaitHandle() method on all non-WaitHandles and this was catching most of
accidental misuse. We no longer do that after removal of userland Awaitables
and it became very easy to cause a segfault. Let's fix it.

Reviewed By: mxw, alexeyt

Differential Revision: D7377537

fbshipit-source-id: 70c6afbc9969ded80ac26362a1fe217f0cd4dad9