Fix loading of SrcLocs from repo.
Summary:
If we got a hit for our unit in the repo, we would have an empty m_sourceLocTab but we would still try to stash it, so things that tried to get SrcLocs would never find any. Only stash the SrcLocs if we actually have them. If we loaded the unit from the repo we won't have them initially, but they'll be lazy-loaded if needed.
Mangle RuntimeOption::RepoDebugInfo into the unit md5, so that if we wrote to the unit to the repo previously with RuntimeOption::RepoDebugInfo=false and we're now running with RuntimeOption::RepoDebugInfo=true we will recompile to get the debug info.
Remove redundant Option::RepoDebugInfo and just use RuntimeOption::RepoDebugInfo, but preserve the behavior that it is on by default _except_ when building a repo.
Finally, fix a bug in hphp/test/run where intermediate .hhas files ending in hhcodegen_output.hhas (from the --run-hh-codegen or --compare-hh-codegen test modes) can be picked up as tests themselves, and clean up a number of intermediate files created in the --hhbbc2 and --compare-hh-codegen test modes.

Reviewed By: markw65

Differential Revision: D7362103

fbshipit-source-id: 33a51247241889d456e6721ff137f4b3cb95556a