Kill Object/Reference getTyped and is helpers
Summary: I noticed these when looking for users of dyn_cast and friends. The getTyped helers are deprecated and totally unused. The is helpers are mentioned in the same docblock as getTyped, the Object one is totally unused, and the Resource one is used in a single place where it's trivial to stop using it. Delete these.

Reviewed By: markw65

Differential Revision: D7219264

fbshipit-source-id: 258fd79cc2795b339b09db45a17ea1a763a04ed9