Failure when checking type-checker todo list was reverting environment
Summary: Fix is a single prime: when executing subtype checks in the "todo" list, the wrong environment was returned in the case that the subtype check failed. This was causing issues with another change (return type inference). I don't have a good repro, but hopefully it's clear that the code was wrong.

Reviewed By: vladima

Differential Revision: D7381929

fbshipit-source-id: 023ac4b32fa99a3ae8108d5d0eb366f3de02f212