HHVM: Change empty string handling.
Summary:
The current logic has the effect of asserting that the first string the callback sees is nonempty. From jemalloc's point of view, emitting an empty string is a valid way to emit no output. This combination breaks on a recent jemalloc stats refactoring.

Instead, we special case empty output strings.

Reviewed By: markw65

Differential Revision: D7374772

fbshipit-source-id: 0039dc6a6ca4acde60cf6e330af70ec9da9cf14a