Add parser error for for-loops with as-expressions
Summary:
When users use as expressions in for-loops, the errors they get are
less-than-ideal.

Given
```
<?hh // strict

class C {
  public function foo(): void {
    for (dict[] as $key => $value) {
        continue;
    }
  }
}
```
we give
```
$ hh_parse --full-fidelity-errors test.php
(6,25)-(6,25) A type specifier is expected here.
(6,28)-(6,28) An expression is expected here.
```
instead, we should
```
$ hh_parse --full-fidelity-errors test.php
(6,5)-(8,5) For loops can not use as-expressions. Did you mean foreach?
(6,25)-(6,25) A type specifier is expected here.
(6,28)-(6,28) An expression is expected here.
```

Reviewed By: arxanas

Differential Revision: D7398823

fbshipit-source-id: e56e2e0240eb7562477afc49b4fb5c7265aa8b4e