set HGPLAIN through env command in hg.ml wrapper
Summary:
ebdfa2 is causing a strange performance regression ...
Here's the diff in question https://our.intern.facebook.com/intern/diff/D7263352/

the relevant change here is setting the environment variable
HGPLAIN=1 through the command Process.exec which delegates to
Unix.create_process_env where there is an environment and
Unix.create_process when there is none.

This is a quick fix to keep the HGPLAIN=1 in the environment
while avoiding the troublesome OCaml function.

Reviewed By: dabek

Differential Revision: D7403606

fbshipit-source-id: 80f6940c325a27b9996a16ec0fd0df49b24bd415