Improve formatting of classish declarations with type parameters
Summary:
The formatting rules for classish declarations need a little more attention. When long lists of type parameters are present, we currently break and nest a little too aggressively.

This change makes the split after the class keyword an independent split (so that we don't break there just because we had to break elsewhere in the declaration), associates the splits before the extends/implements keywords in a new Parental rule (so that if we have a complex declaration, extends and implements lists are clearly visible because they are at the beginning of a line), and uses independent splits for extends/implements lists with only one item, so that we produce this:

```
  extends C<
    Foo,
    Bar,
    Baz,
  >
```

instead of:

```
  extends
    C<
      Foo,
      Bar,
      Baz,
    >
```

Reviewed By: arxanas

Differential Revision: D7355221

fbshipit-source-id: 881121fbc54909fbf5d9090099ed90665d2838d8