Turn off Eval.LoadFilepathFromUnitCache when dumping to hhas.
Summary:
Turn off Eval.LoadFilepathFromUnitCache when dumping to hhas in unit tests. We need it on when running from .hhas files (so that we see the original filename in e.g. fatals) but it is harmful when we are dumping to hhas (because if we have 2 files with identical content at different paths, we compile the first one and stash it in the repo, then the second one will get a hit in the repo and be emitted to hhas with the wrong path). This fixes a class of failures in the diff I'm working on to switch from `-m dumphhas` to `-vEval.DumpHhas`.
I will put up a diff later to do this automatically based on whether we're loading from a .hhas file with a .filepath or not and remove the Eval.LoadFilepathFromUnitCache option, but this is a simpler short-term fix.
Also, don't need -vEval.AllowHhas=1 when dumping to hhas, it only affects behavior when reading .hhas files or inline hhas, so don't put it in dump_hhas_cmd (it's already in the options we run with in --run-hh-codegen and --hhas-round-trip mode).

Reviewed By: swtaarrs

Differential Revision: D7388165

fbshipit-source-id: 05cfab165d90e4495906d6fe07183231bf87532c