Use catch blocks for finally instead of fault funclets
Summary:
This allows us to Await in finally blocks. async-dispose.php was
previously asserting because we suspended an async function with an active
entry in g_context->m_faults, causing a variety of issues. By using a catch
block rather than a fault funclet, the exception is stored in an unnamed local
in the async function and it can be suspended and resumed normally. The only
complication was retaining the automatic chaining done by fault funclets; this
is accomplished with a new ChainFaults bytecode.

It's not currently possible to do a normal await in a finally block, but it can
be done with using. We can think about allowing other awaits in finally blocks
later.

I also noticed a refcounting bug in chainFaultObjects(): it was over-increfing
prev. Fix it by using tvMoveIgnoreRef rather than tvSetIgnoreRef, and add a
test for it.

Reviewed By: markw65

Differential Revision: D7388713

fbshipit-source-id: 5ff7042f4d9df392da5a39e3be902d6f444ae0fc