Conditionally skip 2 quick tests when -vEval.DisableHphpcOpts=1
Summary: We have 2 compiler stress tests that verify that we don't crash when compiling very long expressions composed of additions or string concatenations. The reason we don't normally crash for these inputs is that constant folding reduces them to single constant values before we start emitting. When -vEval.DisableHphpcOpts=1 then the constant folding is disabled and we crash when running out of stack in the EmitterVisitor, but we have some test modes that intentionally do -vEval.DisableHphpcOpts=1. Skip the tests in those modes.

Reviewed By: markw65

Differential Revision: D7410195

fbshipit-source-id: 7e8d5a80e3b90a852c2ca8fb609673aba4f70d00