Augment hh_server with dynamic_view flag
Summary:
This diff adds a ref to Typing_check_service which keeps track of whether dynamic_view is turned on.

When the ref is on, files in the ide will show extra errors. Most of this diff is just threading through the list of ide files to Typing_check_service.

I will make hh_client send a new LSP rpc to hh_server to set this ref, but for now it's set to false by default. I'll add an ability to toggle this in the next diff.

Reviewed By: pittsw

Differential Revision: D7355103

fbshipit-source-id: e2656d90e854fc314642037706510eb49869532d