Use smart constructors in Decl Mode
Summary:
We change underlying implementation of decl mode with smart constructors. Now smart constructors will drop function bodies earlier and this reduces garbage collection overhead.

I replace function body with empty `CompoundStatement` is function doesn't have `yield` keyword in it and with yield expression if it's in the function body somewhere. This way I can pattern match in the lowerer and set `saw_yield` correctly.

In the future diffs I'll try to remove this additional logic introduced in `pFunctionBody` to distinguish between generator function and simple one. But for now I see errors with wrong positioned information if I try to do this in the naive way.

Differential Revision: D7265695

fbshipit-source-id: 5724ce8fe26a25d909cdbd182ac39d0ea0543cbf