Use hhi definitions in test_infer_type and test_server_hover
Summary:
We've been defining builtins as needed for these tests, but it's easier to just use the actual definitions.

Ensuring we have the actual definitions has an impact on typechecking as well--we are now able to properly typecheck Class_const calls on a `classname<ClassVariable>` in the modified server_hover test, since `classname` and `typename` are defined in hhi.

Reviewed By: pittsw

Differential Revision: D7424191

fbshipit-source-id: 25eee144b495c83e6120f234d4197f7808ad2775