Add tpenv to TAST; use in ServerHover to show constraints
Summary:
Currently, the bounds on type parameters seem to be represented only in upper bounds generated when we localize thse parameters. This means that the TAST contains very little information about expressions typed as one of these type parameters--they are annotated with `Tabstract (AKgeneric "T", None)`, with no upper bound for `T`. Without this, it is not possible to look up the members of `T` or meaningfully verify uses of expressions typed `T`.

To address this, this change adds a `tpenv` property to the Tast's EnvAnnotation, which stores the union of the global and local tpenv after localizing generic parameters. When visiting an EnvAnnotation in Tast_visitor, we restore the entirety of the saved `tpenv` to `global_tpenv`. In order to avoid circular dependencies, this change moves the definition of the `tpenv` type to a new file.

This diff exercises these changes by printing constraints in ServerHover.

Reviewed By: jamesjwu

Differential Revision: D7424734

fbshipit-source-id: b5cd19144b79eeb5411df8bc6bf045bb3fce9100