Union lambda environments with containing environment
Summary:
Currently, when Tast_visitor enters a lambda, it uses the `tenv` and `subst` from the lambda's EnvAnnotation directly. This means that we don't make use of any unification which happened after the lambda's definition (which can include, for instance, the growing of `Tunresolved` types). See the added test case for an example.

With this change, we union the `tenv` and `subst` from the lambda with those from the containing environment. When a key occurs in both maps, we use the one from the containing environment.

Reviewed By: andrewjkennedy

Differential Revision: D7184588

fbshipit-source-id: 814fef08f5dbe0ae61a9dec2fe777bd7c440bee9