Typing_print.Full: Improve formatting of delimited lists
Summary:
This change wraps most lists in Typing_print.Full in a Span and a WithRule region using a Parental rule.

The Span has the effect of penalizing adding line breaks within lists which would fit on one line.

The WithRule region ensures that lists break in an all-or-none fashion, producing:

```
dict<
  string,
  LongClassName
>
```

Where, without the WithRule, we would have produced:

```
dict<string,
  LongClassName>
```

Using a Parental rule for the block ensures that when a line break occurs in a nested list, all containing lists break too, so that we produce:

```
dict<
  string,
  dict<
    string,
    LongClassName
  >
>
```

Where, if we had used a Simple rule instead of a Parental rule, we might have produced:

```
dict<string, dict<
  string,
  LongClassName
>>
```

Reviewed By: pittsw

Differential Revision: D7439793

fbshipit-source-id: 97eda631b05b28faf447f546cc1bc73adbfb633e