Add --typed-ast command
Summary:
The --typed-ast command takes a file and outputs a positioned full fidelity parse tree with a 'value' field
containing a position and optional types fields. The base JSON format is the same as that of 'hh_parse --full-fidelity-json'.

The position field contains the start and end position of the tree with:
- all line/column indexes 0 based
- start inclusive
- end exclusive (1 past the end of the range)

The types field contains an array of Typing_print.to_json entries collected from the Typed AST.

Currently the resulting JSON includes type annotations for expressions and class ids.

Plan is to include type annotations for type hints (aka types in symbol declarations) in a later diff.

Position information between the parse trees and the typed asts are incompatible. As such, the lookup of the type information is inefficient. This can be improved once the full fidelity parser is the parser of reference for typing.

Differential Revision: D7425914

fbshipit-source-id: 003e1d050de44577aa14bdcd281af6ebae4dad0a