make inline asm stp instruction work for aarch64/clang
Summary: clang/llvm has multiple issues with the stp instruction.  The one addressed by this diff is that it will not accept the same inline asm syntax for stp that gcc does.  This works around that problem.  A second diff to llvm is necessary to get this fully working - D7433817

Reviewed By: markw65

Differential Revision: D7462020

fbshipit-source-id: 79f954a5f0e25488bcb9f0842e147326642d96d7