Fix --hphpc and --compare-hh-codegen in repo mode.
Summary:
Fix --hphpc when used together with --repo: the --hphpc option was plumbed into the hhvm_cmd function that is used in non-repo modes, but it was not being handled in the hphp_cmd function that is used in repo modes.
Similarly, --compare-hh-codegen together with --repo was comparing the equivalent of --hackc to default, but the default mode is also running hackc. Make hphp_cmd explicitly set the compiler to hphpc or (embedded) hackc when in --compare-hh-codegen mode.

Reviewed By: paulbiss

Differential Revision: D7469326

fbshipit-source-id: 9187ace08f4a13dab899144a87502935efeb09bf