LdClsCtx and LdClsCctx can apply to any 'context'
Summary:
The documentation claims they only apply to the current context class,
and there are a couple of asserts in the simplifier. But nothing
besides the asserts depends on this, DCtxCls already assumes they may
be applied to an unrelated ctx, and if nothing else, the inliner can create
them in the calling function, but applied to the callee's ctx (which can
cause the asserts to fire if the caller isn't a method).

Reviewed By: paulbiss

Differential Revision: D7466857

fbshipit-source-id: c4a0c80ac978583ae52ea311edd8f7e47c5ffba3