Handle systemlib units consistently
Summary:
We stored them in the repo under the name "/:system*", but the
UnitEmitters were generally missing the /:, as were the units
themselves.

In addition, we didn't store them in the unit-cache, so it was
impossible to get hold of a systemlib unit by name (calling lookupUnit
on its name would fail because it didn't have the "/:", and if you
added the "/:" it would create a new unit, rather than returning the
existing one).

Reviewed By: paulbiss

Differential Revision: D7469199

fbshipit-source-id: 7e1f5089e97e886c589fa73ed63338ddf397c94e