Increase cost of RequireClause split, add missing Nest
Summary:
Currently, the split between `require implements` and the interface name is a `Split` without a containing `WithRule` region. Because of the way splits and chunks are associated by Chunk_builder, when the interface name has generic parameters, the split ends up associated with the `Parental` rule governing the splits in the parameter list, so we split after the `implements` keyword unnecessarily:

```
  require implements
  Aaaaaaaaaaaaaaaaaaaaaaaaaaa<
    Bbbbbbbbbbbbbbb,
    Cccccccccccccc,
  >;
```

The interface name is also aligned with the `require implements` keywords because it is not wrapped in a `Nest`.

This change makes the split an independent split (i.e., `SplitWith`), and gives it a high cost (which is renamed from `SimpleMemberSelection` now that it is used in more than one place). The independent rule governing the split ensures that it does not break just because the interface name cannot fit on one line (as a `Parental` rule would). The high cost ensures that we do not produce this formatting:

```
  require implements
    Aaaaaaaaaaaaaaaaaaaaaaaaaaa<Bbbbbbbbbbbbbbb, Cccccccccccccc>;
```

Since without the high cost, we would try to avoid breaking up the `Span`-wrapped parameter list by breaking on the RequireClause's independent split instead. This formatting isn't unreasonable, but I figured breaking up the parameter list was more desirable.

Reviewed By: arxanas

Differential Revision: D7476345

fbshipit-source-id: d344617bbc92fbca471ece677d91d5cddb53cc13