augment do not replace environment when calling
Summary:
This is one potential solution to guard ourselves against future environment
variable issues.

Augmenting the environment, potentially shadowing whatever was underneath is
probably the behavior that most programmers expect when passing in environment
variables.

However, there are a few problems with doing this

1) It's not how exec* works
2) It's not possible to explicitly remove things from the environment anymore
3) naming the parameter "env" does not make it obvious whether we're dealing with
an augmentation or replacement

Potential other solutions:

1) remove the ability to set environment variables in the process api. The alternative
would be setting things like HGBLAME in the master process. That way all the processes
always have the same environment
2) make the API explicit about whether you want to replace the environment or "upsert"
variables

Reviewed By: dabek

Differential Revision: D7406487

fbshipit-source-id: 5dc90796730c8db7fffb0239d425c040ac7993d7