Stop extracting the embedded compiler into /tmp in hphp/test/run
Summary:
Pass the same value of Eval.HackCompilerExtractPath to hphp and hhas dumping steps that we do to hhvm, so that all test steps can use the same single copy of the extracted embedded compiler.
Set Eval.HackCompilerDefault = false in hhbbc; it doesn't support setting runtime options on the command line, but it shouldn't have to compiler anything and this keeps it from extracting the embedded compiler into /tmp (and later removing it).

Reviewed By: oulgen

Differential Revision: D7487560

fbshipit-source-id: 018037d12cba47b5b9e488624ceb17cc749f4383