Split up FuncAnalysis
Summary:
FuncAnalysis was used to store the results of analyze_func. It
contains all the information needed to update the index, and also all
the per-block analysis state needed to optimize the function.

During the analysis phase, we don't need the per-block state, but we
do need to keep one of these for every function in the working set
(which peaks at every function); during the optimization phase, we
need all of it, but we only need to keep one per active thread.

This splits off a FuncAnalysisResult base struct for use during the
analysis phase.

Reviewed By: ricklavoie

Differential Revision: D7493209

fbshipit-source-id: 9e2764dc18e295beb86a1081f60eb2a16dc1a138