Add support for runtime enforcement of safe-pass-by-ref
Summary:
Introduce ThrowOnCallByRefAnnotationMismatch option, which raises a runtime
exception when a safe-pass-by-ref violation is encountered. This is similar to
the already existing WarnOnCallByRefAnnotationMismatch option.

Guaranteed enforcement of safe-pass-by-ref violations will make it possible to
implement a simpler (a single HHBC opcode) and possibly more efficient calling
convention in subsequent diffs.

Reviewed By: markw65

Differential Revision: D7452527

fbshipit-source-id: fc892b95b4329bb2e5ef5c2ea577d35b63eaf396