Fix possible infinite loop in APC
Summary:
When inserting a value into APC, a check is done to see if the value is
recursive (among other things). For objects, we use toArray() to get the values
to recusively check, which for collections does int-like string key
coercion. This means you can "hide" a value in a collection from the circularity
check. However, once we go to actually convert the object, we don't use
toArray(), so we see both values. So, you can hide a recursive value, and then
we'll loop infinitely when trying to convert. Fix this by using asArray() for
collections, which simply just returns the underlying array without any
conversion.

It might be possibly to exploit this with other types of objects as well.

Reviewed By: paulbiss, alexeyt

Differential Revision: D7496142

fbshipit-source-id: beb2138d6943d9d68d6ff28ffbe226454649ecc8