Serialize profile data for standalone jitting
Summary:
This adds an option to serialize the profiling state before running
retranslateAll, and another option to read in that state, and run
retranslateAll during hphp_process_init.

It adds a --jit-serialize mode to hphp/test/run to test it.

This initial version does not attempt to serialize/deserialize
TargetProfile data; but we can run retranslateAll without that, and
I'll add it in follow up diffs.

Reviewed By: swtaarrs, paulbiss

Differential Revision: D7481423

fbshipit-source-id: fa5b78bb2522ecaa21373f5caa78bcf81643b93b