Add proper support for nonnull in HHVM
Summary: As the title says.  Previously `nonnull` was treated as an alias of `mixed`.  This diff tries to ensure that a function argument with the `nonnull` type hint is checked for nullity and that `nonnull` is properly handled by reflection and type structure mechanism (and, in particular, is distinguishable from `mixed`).

Reviewed By: ricklavoie

Differential Revision: D7399943

fbshipit-source-id: 5164e5e2c996a84bcdd46dbcacb2b024f1c9f1f5