Add visibility checks even for non-ConsistentConstruct constructors
Summary:
```
class Base {
  protected function __construct() { ... }
}

class Derived extends Base {
  private function __construct() { ... }
}
```
should be an error, because the child restricts access to a constructor the
parent explicitly made available. This is disallowed for methods and, so far,
only for constructors of classes with the ConsistentConstruct attribute.
Since there are only three cases of this (2 in www and 1 if stdlib in fbcode)
and it causes a Fatal in HHVM, we should just remove these cases and
explicitly disallow this.

Differential Revision: D7504983

fbshipit-source-id: aee2d03c46b937f549546dfe1447e64ef5709ecf