Retain dynamic_view on server restart during LSP
Summary:
This diff adds a --dynamic-view argument to serverArgs, which clientLsp utilizes.

When the server sends an onToggleCoverage command, we cache the toggled coverage result and pass it to the Hack server when the server is restarted in Nuclide.

Slightly related to this diff: There is one other place where the toggleDynamicView could go out of sync: the LspLanguageService itself could crash, instead of clientLsp. But since LspLanguageService is run inside the nuclide server, its crash would result in restarting the client side Nuclide anyways, so this should almost always stay in sync.

Reviewed By: arxanas

Differential Revision: D7509053

fbshipit-source-id: ffceb31056cb4b08b7c94ca12e24219176233828