Fix a trait flattening bug
Summary:
If we flatten a trait method which contains a local function
definition, the flattened method will refer to the local function by
index in the trait's unit. If the class it was flattened into was in a
different unit than the trait, that index will be meaningless, and
things go badly wrong.

We could copy the local function into the target unit, and rewrite its
index, but this situation is very unusual, and we already deal with
the same issue for local classes by preventing the trait from being
flattened; just do the same for local functions.

Reviewed By: ricklavoie

Differential Revision: D7516463

fbshipit-source-id: 0afd20250a56b75c5e665cf61ffd0926f8b6df68