fix garbage in environment use setenv not putenv
Summary:
Somehow, the old code is putting uninitialized garbage in
the environment. This is surprising since we're initializing
everything in the array to zero, strcating stuff into place
and then handing it over to putenv.

putenv however is discouraged anyway and so I changed to code
to use setenv.

I also added an assertion that the type we're getting *really* is
an OCaml string.

Reviewed By: arxanas

Differential Revision: D7536580

fbshipit-source-id: 9c0a13481b6637a2554b344cc87613b6f94a1ac8