Allow unnamed type structures in TypeStructure::resolve
Summary:
Showing the user `TypeStruct$DummyTypeName` in an error message is clowntown. This allows callers of `TypeStructure::resolve` to omit a name, and the error message is adjusted accordingly.

I verified that the `name` field of the constant is not used anywhere else, and we still assert it exists for type constants.

Reviewed By: alexeyt

Differential Revision: D7512446

fbshipit-source-id: 65862feb4b74be262ce1b9978d59a211478aeecb