Improve dependency tracking for "Unbound name" errors
Summary:
See {D6970960} for my previous struggle with this problem. I hoped that I can just pinpoint all the places where we name, but not check a expression, but there seems to be more and more of them (see T27930077 and T27934022, leading to more trouble in comments of T27850330).

This is probably only, like, 60% correct, but still better than 100% incorrect thing that we do now. Consider it a temporary patch to reduce bleeding.

Reviewed By: alexchow

Differential Revision: D7510930

fbshipit-source-id: 0412a74fb74f63f14f1ecaa0dee1a2f89ff5bba3