Run cleanup_for_emit in parallel with make_unit_emitters
Summary:
I noticed that its taking a long time just to run the newly added
cleanup_for_emit, so run it in parallel with make_unit_emitters.

Note that this whole step runs in parallel with writing the unit
emitters to the repo, so previously the time to destroy the index (a
superset of cleanup_for_emit) was hidden behind the time taken to
write to the repo.

Reviewed By: alexeyt

Differential Revision: D7538457

fbshipit-source-id: 2f0b3ccbc1d4a25122b9d743a238739de27795a3