Typecheck tuples in is-expressions
Summary: Ordinarily, we'd just set the type of the variable to be the tuple designated by the hint. But, if we know that the variable is already a tuple of the same length, we can be a bit more intelligent and recursively refine the inner types. This allows us to use the safe instanceof behavior for the elements.

Reviewed By: manzyuk

Differential Revision: D7439468

fbshipit-source-id: 0ad357bd237031965c56a9559ed32ed483c90168