Avoid extra newline in log_duration
Summary:
{D7464038} broke some of the Hack tailing code, which allows
Hack clients to figure out the status based on the last line of the log
file. By adding a newline in log_duration, the clients no longer
recognized the output of log_duration.

I refactored this file a little bit to make the names more descriptive
and to hide the `print_raw` function.

Reviewed By: dabek

Differential Revision: D7556280

fbshipit-source-id: 2270f8d3fe6664e815406bdad38b10a13481f462