Radix map for memory range tracking
Summary:
The radix tree stores the pointer and size of each range
using a tree of power-of-2 sized nodes. Each node represents
a fixed number of bits in the stored pointer. Slots contain
either a pointer to a child node, or the size of an object
starting at that slot's address, and exactly fitting one or
more slots at that tree level.

When ranges have an aligned pointer *and* an aligned size,
such as we have with our 4-buckets-per-power-of-2 size class
scheme, allocations tend stored higher in the tree because
both the pointer and size will be an even number of slots
at intermediate tree levels. Furthermore, when the underlying
allocator subdivides address ranges for small objects, those
objects should cluster nicely in the radix tree.

Reviewed By: alexeyt

Differential Revision: D6871459

fbshipit-source-id: 83cf315d97420050784024096adfafe10f796026