Avoid unnecessary Tunresolved's when intersecting local environments
Summary: When the types of a local variable are physically identical in two branches of a conditional, the type of that local after the conditional should be the common value of the types, not an unresolved containing those types.  In particular, a conditional, should not change the type of `$this` from `<static> as this as Foo` to `<static> as this as ^(Foo)`.

Reviewed By: andrewjkennedy

Differential Revision: D7535880

fbshipit-source-id: e379d5cd9b872a481d133e0bfc23b40831931eef