rpc_persistent now calls back on push
Summary:
Prior to this diff, when you called `ServerCommand.rpc_persistent`, then it would block until the RPC call had responded. It would return (1) the result of the RPC call and (2) all the push messages that had happened while we were waiting for it.

After this diff, it still blocks until the RPC call has responded. But rather than queuing up all the push messages until the end, it now dispatches them immediately to a callback. The callback uses a `state` monad so we can do this in a clean functional way.

Immediate goal is so I can react immediately when I get a `ServerCommandTypes.BUSY_STATUS` push, e.g. by updating the UI straight away.

(Note: if we wanted, we could use this mechanism to support formatting/outline handling even while the server is busy, by passing the client-message-queue fd into the wait statement and responding to messages from there as well, but it'd be pretty gnarly.)

Reviewed By: pittsw

Differential Revision: D7391650

fbshipit-source-id: f39c7fd34b5cab22392a981aee4b7480ca6c50e6