LSP logs server state at time of message arrival
Summary:
For each incoming message from the editor, we already log how long it took us to handle it, and other details.

This diff adds to that log some information about what was the state of the hh_server at the time the message was received. The motive is so, when we review telemetry, we can get better insight into the various causes of slow/absent responses back to the editor.

To get information about hh_server state at the time the message was received is a bit tricky. That's because message-receving happens in a completely separate jsonrpc process, spun off as a daemon, so it doesn't suffer even when the main clientLsp loop is blocked. But clientLsp is the only thing that knows what the server status is, based on the most recent thing it heard from the server.

I didn't want to harm the isolation of jsonrpc. So instead, I'm making clientLsp keep a log of the states it's been in over the past 2 minutes. That way it can look back historically to see what the state was.

Alternatively: we could have tried to log the server state after we'd finished handling the message. But that will by definition never be "typechecking". Or we could log the server state after all previous clientLsp commands had finished, but without knowing whether the server was busy with some work driven by external client requests.

Reviewed By: pittsw

Differential Revision: D7391656

fbshipit-source-id: ba8d2bbd8419287dd33f9365a389af3bc7c5ae03