LSP gives better "denied" messages
Summary:
Prior to this diff, if the server denied a request due to initializing, or being stopped, or blocking all requests due to a rebase, then it just gave back the response RequestCancelled("Server busy") to the editor. (It wasn't user-visible; just went in the logs).

After this diff, the error message now distinguishes between those cases.

Reviewed By: pittsw

Differential Revision: D7391658

fbshipit-source-id: 60e289daa5cecc592c74e1480486f6ba30aa29ba