Untemplatize TargetProfile::data()
Summary:
In the current implementation, each callsite needs to know how it
wants to call TargetProfile::data(), but in practice they're all the
same, except for SwitchProfile which needs a count. Rewrite it so we
always pass in the size of the rds allocation, and always call T::reduce.

This will make it easier to generically serialize/deserialize
TargetProfiles in the next diff.

Reviewed By: swtaarrs

Differential Revision: D7553050

fbshipit-source-id: 7624ebc038174e1934af4fd3471d36f295f57469